Add transactions, read me.
Hi, this is a big commit. I am sorry.
I have added several things in this.

Firstly, accounts now use UUID's instead of Ints.
A Transactions table exists, including a Transactions Class.
The Transactions class is used to formulate a transaction.
A Util folder contains that, and the enums used to define its information.

Do not use the enums or transaction to pass information, use the database for that.
Keep the methods sideeffect-less.

Note: a tryTransaction method has been implemented inside of the
ExampleController.class, please have a look at it. There is no tests for it, yet.
Let me know if there is anything you disagree with, or find odd. :)